+++ mod/racing_plus/PostRender.lua
@@ -67,6 +67,7 @@ function PostRender:Main()
   -- Check for inputs
   PostRender:CheckConsoleInput()
   PostRender:CheckResetInput()
+  PostRender:CheckBombInput()
   PostRender:CheckKnifeDirection()
 
   -- Make Cursed Eye seeded
@@ -134,12 +135,10 @@ function PostRender:DrawStreakText()
   -- Draw the string
   local posGame = g:GridToPos(6, 0) -- Below the top door
   local pos = Isaac.WorldToRenderPosition(posGame)
-  local f = Font()
-  f:Load("font/droid.fnt")
   local color = KColor(1, 1, 1, fade)
   local scale = 1
-  local length = f:GetStringWidthUTF8(g.run.streakText) * scale
-  f:DrawStringScaled(g.run.streakText, pos.X - (length / 2), pos.Y, scale, scale, color, 0, true)
+  local length = g.font:GetStringWidthUTF8(g.run.streakText) * scale
+  g.font:DrawStringScaled(g.run.streakText, pos.X - (length / 2), pos.Y, scale, scale, color, 0, true)
 end
 
 function PostRender:TheLostHealth()
@@ -388,6 +387,41 @@ function PostRender:CheckResetInput()
   else
     -- In speedruns, we want to double tap R to return reset to the same character
     g.run.fastResetFrame = isaacFrameCount
+  end
+end
+
+function PostRender:CheckBombInput()
+  -- Play the sound effect at most once per floor
+  if g.run.playedSad then
+    return
+  end
+
+  -- Check to see if the player has pressed the restart input
+  -- (we check all inputs instead of "player.ControllerIndex" because
+  -- a controller player might be using the keyboard to reset)
+  local pressed = false
+  for i = 0, 3 do -- There are 4 possible inputs/players from 0 to 3
+    if Input.IsActionTriggered(ButtonAction.ACTION_BOMB, i) then -- 8
+      pressed = true
+      break
+    end
+  end
+  if not pressed then
+    g.run.bombKeyPressed = false
+    return
+  end
+
+  if g.run.bombKeyPressed then
+    return
+  end
+  g.run.bombKeyPressed = true
+
+  local bombs = Isaac.FindByType(EntityType.ENTITY_BOMBDROP, -1, -1, false, false) -- 4
+  if #bombs == 0 and
+     g.p:GetNumBombs() == 0 then
+
+    g.sfx:Play(SoundEffect.SOUND_SAD, 1, 0, false, 1) -- ID, Volume, FrameDelay, Loop, Pitch
+    g.run.playedSad = true
   end
 end